refactor: AssetBalanceフックの責務をドメイン配下へ整理
- useAssetBalanceをassetBalance feature配下へ移設
- 汎用useReceiptDataSourceを削除しAssetBalance専用フックを追加
- 関連コンポーネントとテストのimport先を更新
